+++ questions/1/how-to-loop-through-an-array-in-javascript/how-to-loop-through-an-array-in-javascript.html
@@ -5,7 +5,7 @@
     <link rel="icon" type="image/svg+xml" href="/vite.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Recently Active Questions - Stack Overflow Clone</title>
-    <link href="../../src/style.css" rel="stylesheet" />
+    <link href="../../../src/style.css" rel="stylesheet" />
     <link rel="preconnect" href="https://fonts.googleapis.com" />
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
     <link
