+++ src/main.ts
@@ -1,7 +1,6 @@
 import {
   createQuestionElement,
   fetchQuestions,
-  fetchCurrQuestion,
   fetchTags,
   fetchUsers,
   type User,
@@ -25,10 +24,3 @@ questions.forEach((question) => {
 
   questionsContainer.appendChild(questionElement)
 })
-
-const questionTitle = document.querySelectorAll('a')
-questionTitle.forEach((id) =>
-  id.addEventListener('click', () =>
-    console.log(fetchCurrQuestion(parseInt(id.id)))
-  )
-)
